figure out how to get feature gating to work
It was already working, but the build_config file from my source directory is pulled into the python path, with higher precedence than the internals of the package.

I will need to talk to someone who knows more about python packaging how to avoid this.

It works fine as long as the binary is not invoked from the project directory.