make tests work with key from builderB
I think there is still a race condition with uploading the absolutely first signature,
so that one is missing from the data set, and make verification fail with builderB.
Though I have seen this in earlier data only, and not again verified it is the case here.

Marking both keys as trustworthy leads to an endless loop somehow.
This is probably also why the VM tests fail by filling up the disk of the verifier.