split up VM test implementation into files
This makes it easier to inject configuration into tests,
and is done to prepare for exposing more than one test.
In the future this is important for testing the various trust models,
the more pressing issue is encoding both working parameters for larger local tests,
and smaller tests that can run in a GitHub Action.

We also may want to create two versions of the same test, with one filling the cache sequentially,
to not overload storage or RAM of GitHub Actions.
Alternatively we could always execute builders sequentially to fill the cache.