add aterm-based hash and its preimage to signature
We can use that to switch from JSON to ATerm, to be compatible with both nix and lix, as long as lix still supports ca-derivations.
It also makes sense in general to do this until we potentially create our own format, or ca-derivations are stabilized differently,
because with things like dynamic derivations and experimental features in general,
the JSON representation is also unstable between different versions of nix.