try ci run with less strict trust model
This will probably fail on the large test anyways
because there are some things in the closure,
which we are not pre-fetching successfully.

Additionally I think I saw the larger test fail
locally as well, and I need still need to look into
what's wrong there. It would be good to have these
running in CI.