fix cli signing tests
The reason why these tests did not work in the sandbox was the
existance check that was done by click. This would always fail
unexpectedly, and so by mocking the innards of this check,
we finally get to run those tests in the sandbox.
Unfortunateley that requires mocking quite a bit of stuff
right now.